fix: move content helpers into site-config
